+++ web/src/app/utils.js
@@ -94,7 +94,6 @@ export const unmatchedTags = (tags) => {
     else return tags.filter(tag => !(tag in emojis));
 }
 
-
 export const maybeWithBasicAuth = (headers, user) => {
     if (user) {
         headers['Authorization'] = `Basic ${encodeBase64(`${user.username}:${user.password}`)}`;
@@ -240,4 +239,13 @@ export async function* fetchLinesIterator(fileURL, headers) {
     if (startIndex < chunk.length) {
         yield chunk.substr(startIndex); // last line didn't end in a newline char
     }
+}
+
+export const randomAlphanumericString = (len) => {
+    const alphabet = "abcdefghijklmnopqrstuvwxyz0123456789";
+    let id = "";
+    for (let i = 0; i < len; i++) {
+        id += alphabet[(Math.random() * alphabet.length) | 0];
+    }
+    return id;
 }